pyproject.toml
[project]
name = "ProVoice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "carla>=0.9.15",
    "dash>=3.3.0",
    "dash-bootstrap-components>=2.0.4",
    "flask>=3.1.2",
    "matplotlib>=3.10.7",
    "mediapipe>=0.10.14",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "pygame>=2.6.1",
    "yolov5-deepsort-driverdistracted-driving-behavior-detection>=0.1.0",
    "keras>=3.12.0",
    "pandas-stubs==2.3.3.251201",
    "joblib>=1.5.2",
    "scikit-learn==1.7.0",
    "mmrphys",
    "pyttsx3>=2.99",
    "asyncio>=4.0.0",
    "fastapi>=0.124.4",
    "python-socketio>=5.15.0",
    "uvicorn>=0.38.0",
]

[tool.uv]
package = true
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]
mmrphys = { git = "https://github.com/LukasWestholt/MMRPhys" }
yolov5-deepsort-driverdistracted-driving-behavior-detection = { git = "https://github.com/LukasWestholt/Yolov5-deepsort-driverDistracted-driving-behavior-detection" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[project.scripts]
provoice = "ProVoice:main"
drive = "drive:main"
